requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.3
alabaster==0.7.12
alembic==1.4.3
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
appdirs==1.4.4
APScheduler==3.6.3
argh==0.26.2
asgiref==3.2.10
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.8
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.9.3
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
boto3==1.16.4
botocore==1.19.4
Bottleneck==1.3.2
branca==0.4.0
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.4
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
comtypes==1.1.7
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.11.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.11.0
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
fbmessenger==6.0.0
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
folium==0.10.1
fsspec==0.6.2
future==0.18.2
gast==0.2.2
gevent==1.4.0
gitdb==4.0.5
GitPython==3.1.11
glob2==0.7
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14
greenlet==0.4.15
grpcio==1.28.1
guietta==0.3.5
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
h2o==3.30.0.3
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2020.10.20
html5lib==1.0.1
httpcore==0.3.0
httplib2==0.18.1
httptools==0.1.1
httpx==0.9.3
humanfriendly==8.2
hyperframe==5.2.0
hypothesis==5.5.4
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.5.0
imutils==0.5.3
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.15.1
Js2Py==0.70
json5==0.9.1
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyter-server==0.1.1
jupyterlab==1.2.6
jupyterlab-pygments==0.1.1
jupyterlab-server==1.0.6
kafka-python==1.4.7
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
librosa==0.8.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Mako==1.1.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mattermostwrapper==2.2
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.7.6
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==7.0.0
pipwin==0.5.0
pkginfo==1.5.0.1
plotly==4.11.0
pluggy==0.13.1
ply==3.11
pooch==1.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.4
psutil==5.6.7
psycopg2==2.8.5
psycopg2-binary==2.8.6
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.5.2
pyjsparser==2.7.1
PyJWT==1.7.1
pykwalify==1.7.0
pylint==2.4.4
pymongo==3.8.0
PyNaCl==1.3.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypiwin32==223
PyPrind==2.11.2
pyreadline==2.1
pyrsistent==0.15.7
PySide2==5.15.0
pySmartDL==1.3.4
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
python-socketio==4.5.1
python-telegram-bot==12.8
pyttsx3==2.71
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
questionary==1.5.2
rasa==1.10.16
rasa-sdk==1.10.3
rasa-x==0.32.2
redis==3.5.3
regex==2020.6.8
requests==2.24.0
requests-async==0.5.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
retrying==1.3.3
rfc3986==1.4.0
rocketchat-API==1.3.1
rope==0.16.0
rsa==4.0
Rtree==0.9.3
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sanic==19.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.3.2
Sanic-Plugins-Framework==0.9.4.post1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentry-sdk==0.19.1
shiboken2==5.15.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==3.0.4
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
SoundFile==0.10.3.post1
soupsieve==1.9.5
SpeechRecognition==3.8.1
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.20
sqlparse==0.3.1
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tabulate==0.8.7
tblib==1.6.0
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.1.2
tensorflow-addons==0.7.1
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
tensorflow-probability==0.9.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
twilio==6.26.3
typeguard==2.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
voila==0.1.21
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
webexteamssdk==1.3
websockets==8.1
Werkzeug==1.0.0
whitenoise==5.1.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xgboost==1.1.1
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.1
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
yarl==1.5.1
yellowbrick==1.1
zict==1.0.0
zipp==2.2.0
